+++ terraform/environments/examples/org_accounts/outputs.tf
@@ -1,0 +1,4 @@
+output org_accounts {
+  value = {for k, v in module.org_accounts : k => v.organizations_account}
+  // map(object)
+}