+++ terraform/environments/examples/org_policies/outputs.tf
@@ -1,0 +1,4 @@
+output org_policies {
+  value = {for k, v in module.org_policies : k => v.organizations_policy}
+  // map(object)
+}