+++ terraform/modules/org_account/local.tf
@@ -3,7 +3,7 @@ locals {
     var.base_aws_tags,
     {
       Name                      = var.name
-      Terraform_Module          = "aws_org_account"
+      Terraform_Module          = "org_account"
     }
   )
 }