+++ terraform/modules/org_account/variables.tf
@@ -3,4 +3,4 @@ variable name {type = string}
 variable email {type = string}
 variable iam_user_access_to_billing {}
 variable role_name {type = string}
-variable service_control_policy_id {type = string}
+//variable service_control_policy_id {type = string}