+++ terraform/modules/org_policy/outputs.tf
@@ -1,0 +1,4 @@
+output organizations_policy {
+  value       = aws_organizations_policy.default
+  // object
+}